Combin form presenter with form
